+++ lib/rubsh.bash
@@ -228,11 +228,13 @@ __ary_to_str () {
 }
 
 __dispatch () {
-  local method=${1:-to_s}; shift
+  local method=${1:-.to_s}; shift
   local receiver=${FUNCNAME[1]}
   local class=${__classh[$receiver]}
   local statement
 
+  [[ $method == '.'* ]] || return
+  method=${method#.}
   while [[ ${__method_classesh[$method]} != *" $class "* ]]; do
     [[ " ${!__superh[*]} " == *" $class "* && -n ${__superh[$class]} ]] || return
     class=${__superh[$class]}